requirements.txt
google-generativeai==0.8.3
mem0ai==0.0.11
python-dotenv==1.0.0
